+++ src/components/UI/DropdownMenu/types.ts
@@ -1,0 +1,7 @@
+import type { ReactNode } from 'react';
+
+export type DropdownMenuProps = {
+  title: ReactNode;
+  children: ReactNode;
+  menuHorizontalPosition?: 'left' | 'right';
+};
